Master password will not be saved anywhere

+++ utils/files.js
@@ -1,42 +1,47 @@
 const fs = require('fs');
 const path = require('path');
-const { homedir, EOL } = require('os');
+const { homedir } = require('os');
 const { hash, decrypt, encrypt } = require('./encryption');
+const prompt = require('prompt-sync')({ sigint: true });
 
 const configFile = path.join(homedir(), 'vault-config');
+const defaultVault = path.join(homedir(), 'vault');
 
 const exists = fs.existsSync(configFile) && fs.statSync(configFile).size != 0;
 
+let master, config, masterHash, iv, key;
+
 if (!exists) {
-	const prompt = require('prompt-sync')();
-	const password = prompt('Enter a master password: ', { echo: '*' });
-	const data = hash(password) + EOL;
-	fs.writeFileSync(configFile, data);
+	master = hash(prompt('enter a master password: ', { echo: '*' }));
+	masterHash = hash(master, 'sha512', hash(master, 'md5'));
+	[iv, key] = [masterHash.slice(0, 16), masterHash.slice(16, 48)];
+	config = { file: 'vault-config' };
+
+	fs.writeFileSync(configFile, encrypt(JSON.stringify(config), key, iv));
 }
 
-const config = fs.readFileSync(configFile, 'utf-8');
-let [master, customVaultFile] = config.split(EOL);
-let masterHash = hash(master, 'sha512', hash(master, 'md5'));
-let [iv, key] = [masterHash.slice(0, 16), masterHash.slice(16, 48)];
-let customVault = decrypt(customVaultFile, key, iv);
+async function load(reset = false) {
+	if (reset || (exists && !master)) {
+		master = hash(prompt('password: ', { echo: '*' }));
+		masterHash = hash(master, 'sha512', hash(master, 'md5'));
+		[iv, key] = [masterHash.slice(0, 16), masterHash.slice(16, 48)];
+		config = JSON.parse(decrypt(fs.readFileSync(configFile, 'utf-8'), key, iv)) || {};
+	}
 
-const vaultFilePath = customVault || path.join(homedir(), 'vault');
-const vaultFile = path.join(vaultFilePath, fs.existsSync(vaultFilePath) && fs.lstatSync(vaultFilePath).isDirectory() ? 'vault' : '');
+	if (!(config && typeof config == 'object' && config.file == 'vault-config')) {
+		console.log('wrong password');
+		return load(true);
+	}
 
-async function load() {
-	if (!exists) fs.unlinkSync(vaultFile);
+	const vaultFilePath = getVaultPath();
+	const vaultFile = path.join(vaultFilePath, fs.existsSync(vaultFilePath) && fs.lstatSync(vaultFilePath).isDirectory() ? 'vault' : '');
+
 	if (!fs.existsSync(vaultFile)) fs.writeFileSync(vaultFile, '');
 
-	let data = {};
-	try {
-		data = JSON.parse(decrypt(fs.readFileSync(vaultFile, 'utf-8'), key, iv)) || {};
-	} catch {
-		//
-	}
-	return data;
+	return JSON.parse(decrypt(fs.readFileSync(vaultFile, 'utf-8'), key, iv)) || {};
 }
 
-function save(data, path = vaultFile) {
+function save(data, path = getVaultPath()) {
 	fs.writeFileSync(path, encrypt(JSON.stringify(data), key, iv));
 }
 
@@ -47,35 +52,35 @@ async function saveData(path, afterLoad) {
 	});
 }
 
-function saveConfig({ masterPassword = master, path = vaultFilePath }) {
-	fs.writeFileSync(configFile, masterPassword + EOL + encrypt(path, key, iv));
+function saveConfig(path = getVaultPath()) {
+	fs.writeFileSync(configFile, encrypt(JSON.stringify({ file: 'vault-config', path }), key, iv));
 }
 
 function changePath(newPath) {
 	newPath = path.join(newPath, fs.existsSync(newPath) && fs.lstatSync(newPath).isDirectory() ? 'vault' : '');
-	const log = () => console.log(`new vault path: ${newPath}`);
 
-	if (newPath.toLowerCase() != vaultFile.toLowerCase()) {
-		saveConfig({ path: newPath });
-		return saveData(newPath).then(() => {
-			fs.unlinkSync(vaultFile);
-			log();
+	if (newPath.toLowerCase() != getVaultPath().toLowerCase()) {
+		saveConfig(newPath);
+		saveData(newPath).then(() => {
+			fs.unlinkSync(getVaultPath());
+			console.log(`new vault path: ${newPath}`);
 		});
+		return;
 	}
-	log();
+	console.log(`vault path is already ${newPath}`);
 }
 
 function changePassword(password) {
-	saveData(vaultFile, () => {
+	saveData(getVaultPath(), () => {
 		master = hash(password);
 		masterHash = hash(master, 'sha512', hash(master, 'md5'));
 		[iv, key] = [masterHash.slice(0, 16), masterHash.slice(16, 48)];
-		saveConfig({});
+		saveConfig();
 	});
 }
 
 function clearVault() {
-	fs.unlinkSync(vaultFile);
+	fs.unlinkSync(getVaultPath());
 }
 
 function checkFile(filename) {
@@ -88,4 +93,8 @@ function exportFile(filename, buffer) {
 	console.log('entry exported to', filename);
 }
 
-module.exports = { load, save, changePath, changePassword, clearVault, checkFile, exportFile, vaultFile, exists, master };
+function getVaultPath() {
+	return config.path || defaultVault;
+}
+
+module.exports = { load, save, changePath, changePassword, clearVault, checkFile, exportFile, getVaultPath };
